src/requirements.txt
boto3
twilio
botocore
requests
python-dotenv